Refractoring

+++ 0x00-python_variable_annotations/102-type_checking.py
@@ -4,10 +4,10 @@
 from typing import Tuple, List
 
 
-def zoom_array(lst: List[int], factor: float = 2) -> List[int]:
+def zoom_array(lst: List[int], factor: int = 2) -> List[int]:
     zoomed_in: List[int] = [
         item for item in lst
-        for i in range(int(factor))
+        for i in range(factor)
     ]
     return zoomed_in
 
@@ -16,4 +16,4 @@ array = [12, 72, 91]
 
 zoom_2x = zoom_array(array)
 
-zoom_3x = zoom_array(array, 3.0)
+zoom_3x = zoom_array(array, 3)
